corrected the isAccepted table accorrding to RFC 1630
The '/' character is stop to escape due to this commit ca80cbb.
This cause some browser cannot redirect to the original requested page.

Actually, this table isn't correct for lot of characters, so I also
correct it.

'&' is a valid character of xalphas.
However, we still need to escape it, otherwise we can't distinguish it
is a part of the `redir` or a separator.