corrected the isAccepted table according to RFC 1630
The '/' character was used to stop escaped characters due to commit
ca80cbb.
This caused some browsers to not to redirect to the requested page.

Actually, the table is not correct for a lot of characters. This is also
fixed.

'&' is a valid character of xalphas.
However, we still need to escape it as we otherwise can't distinguish if it
is part of `redir` or if it is a separator.