stop using multiple threads to handle HTTP requests
LibHTTPD doesn't support multi-threading at all.  Some static variables,
for example `buf` in `httpdReadRequest`, is shared by all threads.

These cause a lot of unexpected behaviors and sgefaults.
However there are also some patches trying to solve it:
  http://dev.wifidog.org/ticket/471 .

Those patches reduce the chance that cause NoDogSplash to crash, but
the bug still exists.

There are two ways to solve it.  One is to remove the `static` modifier and
to bring more overhead to each request.  Another solution is to not use
threads at all.

I prefer the later one.  Since all requests can be handled very quickly,
I believe that won't increase the response time even if a lot of users
try to access it at the same time.