Update infoskel.html
Often, a client will connect to NDS but not authenticate for a while. When they do, very often the token will have timed out and they are served up infoskel.html.
This is neither mobile or client friendly and leaves them in limbo.
Here is a new version with a button for them to continue. Tested on Apple, Windows, Android and various desktops/laptops.