config: allow gw_address to be an FQDN
This is useful, for instance, if you want to replace an IP address with a
local gateway domain name.  Note that the local resolver MUST return the
gateway address for this domain and we still need the IP address in order to
bind the daemon and install iptables rules.

This actually makes (future) IPv6 support slightly easier, since the
resolver can return both an A record and an AAAA record.

It may be worth adding a similar option to FAS, so that we can allow the IP
but redirect to the FQDN.  This would more easily enable the remote FAS to
be https.