FAS: Introduce hash_id mode. Enhances security and mitigates issues accessing ndsctl remotely to obtain the client token.
In addition, no additional packages are required allowing legacy low flash/ram devices to be used.

For option fas_secure_enabled '1', NDS adds the hash of the client token (hid) to the query string.
FAS concatenates hid and the fas key and hashes the result.
This new hash is added to the query string returned to NDS instead of the client token.
NDS then compares this hash with one it calculates itself from the client token and the pre-shared key.

For option fas_secure_enabled '2', the FAS can return either the client token in clear text
or can return the concatenated hash as for fas_secure_enabled '1', NDS will detect which is used.

Signed-off-by: Rob White <rob@blue-wave.net>