BinAuth: Prevent ndsctl from running if called from a Binauth script. See issue #503
If ndsctl is called within a Binauth script it deadlocks as NDS is in
the middle of its client authentication logic when BinAuth is called.
It should not be necessary to run ndsctl from Binauth anyway as it does not
provide any useful post authentation functionality.

This commit prevents ndsctl from being run using a modified version of
the existing ndsctl.lock logic.

Signed-off-by: Rob White <rob@blue-wave.net>