Force minumum version of libmicrohttpd (MHD)
See issue #515

The setting in main.c works only if NDS is compiled and installed in place.

If creating a package however, microhttpd.h is embedded as an include
leaving no way for NDS to discover the installed MHD version.

This commit uses the NDS init.d script on Openwrt to determine the
installed MHD version.

opkg cannot be used to determine the MHD version as on installing (with opkg)
an opkg lock error is generated.

If in the future, NDS was to be added to other Linux official package repo's
(eg Debian), similar installed version detection would have to be done.


Signed-off-by: Rob White <rob@blue-wave.net>