main: initialize the webserver last
Prevents a race condition, when running the webserver
before initializing the firewall and allowing clients already
to register.
With the state file, this happens much more likely, because
depending on the system, re-adding all clients take a long time.
Most likely because of the many forks of iptables.
On small embedded systems, e.g. the archer c6 v3 it takes for ~160 clients
to load ~10 seconds.